Moving route_summary_cache to private header. Out-lined methods.
